Struct yksom::vm::core::VM

pub struct VM {
    pub block_cls: Val,
    pub block2_cls: Val,
    pub block3_cls: Val,
    pub bool_cls: Val,
    pub cls_cls: Val,
    pub double_cls: Val,
    pub false_cls: Val,
    pub int_cls: Val,
    pub nil_cls: Val,
    pub obj_cls: Val,
    pub str_cls: Val,
    pub system_cls: Val,
    pub true_cls: Val,
    pub false_: Val,
    pub nil: Val,
    pub system: Val,
    pub true_: Val,
    // some fields omitted
}

Fields
block_cls: Valblock2_cls: Valblock3_cls: Valbool_cls: Valcls_cls: Valdouble_cls: Valfalse_cls: Valint_cls: Valnil_cls: Valobj_cls: Valstr_cls: Valsystem_cls: Valtrue_cls: Valfalse_: Valnil: Valsystem: Valtrue_: Val
Methods

impl VM

pub fn new(classpath: Vec<String>) -> Self

pub fn compile(&self, path: &Path, inst_vars_allowed: bool) -> Val

Compile the file at path. inst_vars_allowed should be set to false only for those builtin classes which do not lead to run-time instances of Inst.

pub fn error(&self, error: &str) -> !

Inform the user of the error string error and then exit.

pub fn send(
&self,
rcv: Val,
msg: &str,
args: Vec<Val>
) -> Result<Val, Box<VMError>>

Send the message msg to the receiver rcv with arguments args.