Trait yksom::vm::objects::Obj

pub trait Obj: Debug + GcLayout {
    fn dyn_objtype(&self) -> ObjType;

    fn get_class(&self, vm: &VM) -> Val;

    fn to_strval(&self, _: &VM) -> Result<Val, Box<VMError>> { ... }

    fn add(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn sub(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn mul(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn div(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn shl(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn ref_equals(&self, vm: &VM, other: Val) -> Result<Val, Box<VMError>> { ... }

    fn equals(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn not_equals(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn greater_than(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn greater_than_equals(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn less_than(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn less_than_equals(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>> { ... }
}

The main SOM Object trait. Notice that code should almost never call these functions directly: you should instead call the equivalent function in the Val struct.

Required methods

fn dyn_objtype(&self) -> ObjType

What ObjType does this Val represent?

fn get_class(&self, vm: &VM) -> Val

What class is this object an instance of?

Loading content...
Provided methods

fn to_strval(&self, _: &VM) -> Result<Val, Box<VMError>>

Convert this object to a Val that represents a SOM string.

fn add(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>>

Produce a new Val which adds other to this.

fn sub(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>>

Produce a new Val which subtracts other from this.

fn mul(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>>

Produce a new Val which multiplies other to this.

fn div(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>>

Produce a new Val which divides other from this.

fn shl(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>>

Produce a new Val which shifts self other bits to the left.

fn ref_equals(&self, vm: &VM, other: Val) -> Result<Val, Box<VMError>>

Is this Val reference equality equal to other? Only number types are likely to want to override this.

fn equals(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>>

Does this Val equal other?

fn not_equals(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>>

Does this Val not equal other?

fn greater_than(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>>

Is this Val greater than other?

fn greater_than_equals(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>>

Is this Val greater than or equal to other?

fn less_than(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>>

Is this Val less than other?

fn less_than_equals(&self, _: &VM, _: Val) -> Result<Val, Box<VMError>>

Is this Val less than or equal to other?

Loading content...