Module yksom::vm::val

Tagged pointer support.

Structs

Val
The core struct representing values in the language runtime: boxed and unboxed values are hidden behind this, such that they can be treated in exactly the same way.

Enums

ValKind

Constants

BITSIZE

INT_BITMASK

TAG_BITMASK

TAG_BITSIZE

Traits

NotUnboxable
Objects which impl this trait guarantee that they can only ever be stored boxed. Implementing this trait on objects which can be stored unboxed leads to undefined behaviour.